# [tests] Reduce the number of cBench validation datasets for CI, 7.2x speedup.