@ChrisCummins 2022-06-13
Fixed the test regression.

Using CI/test-llvm-env-linux (3.6) as the benchmark:

- [baseline](https://github.com/facebookresearch/CompilerGym/runs/6868239531?check_suite_focus=true): 55:44
- [after](https://github.com/facebookresearch/CompilerGym/runs/6869148524?check_suite_focus=true): 12:02

**4.6x speedup**.

Cheers,
Chris