@mostafaelhoushi 2022-06-13
Hmmm.. test-linux-cmake is still taking over an hour?