@ChrisCummins 2022-06-13
> Hmmm.. test-linux-cmake is still taking over an hour?

Yeah, not sure what is going on there, my bad! Opened #712 to investigate.

Cheers,
Chris